break into components
